Useless casts removed

+++ compiler/documentator/lateq.cpp
@@ -77,15 +77,15 @@ void Lateq::println(ostream& docout)
     /* 1. Make titles of sub-sets of formulas. */
     string suchthat = gGlobal->gDocMathStringMap["suchthat"];
 
-    string sInputs = makeItemTitle((const unsigned int)fInputSigsFormulas.size(), "inputsigtitle") +
+    string sInputs = makeItemTitle(fInputSigsFormulas.size(), "inputsigtitle") +
                      makeSignamesList(fInputSigsFormulas, "");
-    string sOutputs = makeItemTitle((const unsigned int)fOutputSigsFormulas.size(), "outputsigtitle") +
+    string sOutputs = makeItemTitle(fOutputSigsFormulas.size(), "outputsigtitle") +
                       makeSignamesList(fOutputSigsFormulas, suchthat);
-    string sConstants = makeItemTitle((const unsigned int)fConstSigsFormulas.size(), "constsigtitle") +
+    string sConstants = makeItemTitle(fConstSigsFormulas.size(), "constsigtitle") +
                         makeSignamesList(fConstSigsFormulas, suchthat);
 
     vector<list<string> > UISignamesVector = makeUISignamesVector(fUISigsFormulas);
-    string                sUIElements      = makeItemTitle((const unsigned int)fUISigsFormulas.size(), "uisigtitle") +
+    string                sUIElements      = makeItemTitle(fUISigsFormulas.size(), "uisigtitle") +
                          makeSignamesList(UISignamesVector, suchthat);
 
     unsigned int internalSigsCount =
